Changed some spectra buffer types from "unsignded long" to "int", because this is used internally.

+++ src/specio.c
@@ -10,7 +10,7 @@
 
 
 
-int load_spec(const char *name, unsigned long *buf, int num) {
+int load_spec(const char *name, int *buf, int num) {
 
   int n;
   MFILE *mat;
@@ -22,7 +22,7 @@ int load_spec(const char *name, unsigned long *buf, int num) {
   return n;
 }
 
-int save_spec(const char *name, unsigned long *buf, int num) {
+int save_spec(const char *name, int *buf, int num) {
 
   int n;
   MFILE *mat;
